Major changes.
Modals are used prior to adding a plan or layer to collect some
additional metadata.

A SimMap template directory provides example maps. The map file from
the choosen template is applied during SimMap creation.

When an Agency is created a cooresponding group is automatically created
and given Contributor, Editor status.